requirements.txt
torch
wandb
numpy
scikit-learn
pandas
PyYAML
